Pylint fixes + Re-factor code + Cleanup
- Replace module: json -> ujson
- Split:          mockdata.py -> lastfm_session.py + lastfm_token.py + lastfm_user.py
- Split:          test_mockdata.py -> test_lastfm_session.py + test_lastfm_token.py + test_lastfm_user.py
- Rename:         scrobble_listens() -> record_listens()
- Replace:        type() -> isinstance()
- Remove unused modules, Pylint fixes etc.
